Make keepasshttpc.py package-external executable
If you copy the keepasshttpc folder into your (per user) site-packages
directory and keepasshttpc.py (with whatever name you prefer) into a
directory on your PATH, you can call it as a command line client for
a keepasshttpc. For now, it has to be accessible on localhost:19455.
Refer to the source for documentation on which commands are implemented
(or just try them).